fixed two bugs with prototext format
The first bug was in InitUnsharedWeightsNet. Proto var was of type
ostringstream, which converted the bool bias_term into an int. I
wrote an inline conditional to convert the term into a proper
string. This allows backwards compatibility with earlier prototext
versions.

The second bug was in the syntax used by @mtamburrano for repeated
bool parameters, assigned to the propagate_down parameter. The style
used of e.g. propagate_down: [true,true] does not work with earlier
prototext versions. New style works for all versions.